add test and cleanup
